PreactBaseElement: extract collectProps
